benchmark/inf1_requirements.txt
transformers==4.25.1
protobuf==3.20.1
datasets 
pillow==9.4.0

# Install packages from repos
--extra-index-url https://pip.repos.neuron.amazonaws.com
torch-neuron
neuron-cc[tensorflow]
torchvision